refactor : 유틸 함수 분리에 따른 import 수정
